change global structure, the JPanel extended class now contains the game itself, and the JFrame extended class simply contains the window and handles keyEvents
