+++ MyWindow.java
@@ -1,5 +1,4 @@
 
-import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Toolkit;
 import java.awt.event.KeyEvent;
@@ -9,20 +8,12 @@ import javax.swing.JFrame;
 
 /**class MyWindow extends JFrame<p>
  * Gere la fenetre et les saisies clavier
- * et tout le jeu en soit
 */
 public class MyWindow extends JFrame {
 	private static final long serialVersionUID = 1L;
 
+	/**Contient le jeu en soit */
 	private Board board = new Board();
-
-	/**Booleen, true si le jeu est en cours */
-	private boolean isPlaying = false;
-
-	/**Personnage rouge (initialement a gauche) */
-	private Character characterRed;
-	/**Personnage bleu (initialement a droite) */
-	private Character characterBlue;
 
 
 	public MyWindow() {
@@ -30,51 +21,14 @@ public class MyWindow extends JFrame {
 
 		this.setContentPane(board);
 
-		// ICI : 
+		// ICI a faire plus tard: 
 		// On lance le menu, avec 3 boutons : Play Options Quit
 		// Dans les Options on peut modifier les touches claviers des deux joueurs
 		// Le bouton Play appel les fonctions initGame() et startGame();
 		// Pour le moment on lance juste le jeu
 		
-		initGame();
-		startGame();
-	}
-
-
-	/**Le jeu lui meme (la boucle while true) */
-	public void startGame() {
-		this.isPlaying = true;
-
-		while (this.isPlaying) {
-
-			updateWindow();
-
-			sleep(200);
-		}
-	}
-
-
-	/**Initialise le jeu, creer les deux joueurs avec leurs touches claviers associees serialisees, la ArrayList d'objets */
-	public void initGame() {
-		// On charge les objets (sans image) tout est fonctionnel
-		// Les fonctions d'affichage s'occupent dafficher les images si elles existent, des carres sinon
-		
-		characterRed = new Character(board.getPrimaryXcoordLeft(), board.getGroundLevelYcoord() - board.getCharacterHeight(), Color.red);
-		characterBlue = new Character(board.getPrimaryXcoordRight(), board.getGroundLevelYcoord() - board.getCharacterHeight(), Color.red);
-		
-		// On charge les images, et on les met dans les objets (null si elles n'ont pas reussi)
-		loadAndSetAllImages();
-	}
-
-
-	/**Charge toutes les images du jeu et les ajoute aux objets */
-	public void loadAndSetAllImages() {
-
-	}
-
-
-	public void updateWindow() {
-		repaint();
+		board.initGame();
+		board.startGame();
 	}
 
 
@@ -101,17 +55,6 @@ public class MyWindow extends JFrame {
 		// this.setUndecorated(true);
 
 		this.setVisible(true);
-	}
-
-
-	/**Delay */
-	public void sleep(int time) {
-		try {
-			Thread.sleep(time);
-		} catch (InterruptedException e) {
-			Thread.currentThread().interrupt();
-			e.printStackTrace();
-		}
 	}
